+++ execution_engine_testing/cargo_casper/src/contract_package.rs
@@ -7,14 +7,10 @@ use once_cell::sync::Lazy;
 
 use crate::{
     common::{self, PATCH_SECTION},
-    erc20, simple, ProjectKind, ARGS,
+    simple, ARGS,
 };
 
-static PACKAGE_NAME: Lazy<&'static str> = Lazy::new(|| match ARGS.project_kind() {
-    ProjectKind::Simple => "contract",
-    ProjectKind::Erc20 => "erc20-token",
-});
-
+static PACKAGE_NAME: Lazy<&'static str> = Lazy::new(|| "contract");
 static CONTRACT_PACKAGE_ROOT: Lazy<PathBuf> =
     Lazy::new(|| ARGS.root_path().join(PACKAGE_NAME.replace("-", "_")));
 static CARGO_TOML: Lazy<PathBuf> = Lazy::new(|| CONTRACT_PACKAGE_ROOT.join("Cargo.toml"));
@@ -47,10 +43,7 @@ lto = true
 
 {}"#,
         *PACKAGE_NAME,
-        match ARGS.project_kind() {
-            ProjectKind::Simple => &*simple::CONTRACT_DEPENDENCIES,
-            ProjectKind::Erc20 => &*erc20::CONTRACT_DEPENDENCIES,
-        },
+        &*simple::CONTRACT_DEPENDENCIES,
         PACKAGE_NAME.replace("-", "_"),
         &*PATCH_SECTION
     )
@@ -60,11 +53,8 @@ pub fn create() {
     // Create "<PACKAGE_NAME>/src" folder and write "main.rs" inside.
     let src_folder = MAIN_RS.parent().expect("should have parent");
     common::create_dir_all(src_folder);
-    let main_rs_contents = match ARGS.project_kind() {
-        ProjectKind::Simple => simple::MAIN_RS_CONTENTS,
-        ProjectKind::Erc20 => erc20::MAIN_RS_CONTENTS,
-    };
-    common::write_file(&*MAIN_RS, main_rs_contents);
+
+    common::write_file(&*MAIN_RS, simple::MAIN_RS_CONTENTS);
 
     // Create "<PACKAGE_NAME>/.cargo" folder and write "config.toml" inside.
     let config_folder = CONFIG_TOML.parent().expect("should have parent");